Refactor: Categorization QueryDsl 적용 #91
