+++ src/main/java/com/nhnacademy/marketgg/server/service/impl/DefaultCategorizationService.java
@@ -1,7 +1,7 @@
 package com.nhnacademy.marketgg.server.service.impl;
 
 import com.nhnacademy.marketgg.server.dto.response.CategorizationRetrieveResponse;
-import com.nhnacademy.marketgg.server.repository.CategorizationRepository;
+import com.nhnacademy.marketgg.server.repository.categorization.CategorizationRepository;
 import com.nhnacademy.marketgg.server.service.CategorizationService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
