+++ src/main/java/com/nhnacademy/marketgg/server/service/impl/DefaultCategoryService.java
@@ -7,7 +7,7 @@ import com.nhnacademy.marketgg.server.entity.Categorization;
 import com.nhnacademy.marketgg.server.entity.Category;
 import com.nhnacademy.marketgg.server.exception.categorization.CategorizationNotFoundException;
 import com.nhnacademy.marketgg.server.exception.category.CategoryNotFoundException;
-import com.nhnacademy.marketgg.server.repository.CategorizationRepository;
+import com.nhnacademy.marketgg.server.repository.categorization.CategorizationRepository;
 import com.nhnacademy.marketgg.server.repository.CategoryRepository;
 import com.nhnacademy.marketgg.server.service.CategoryService;
 import lombok.RequiredArgsConstructor;
