+++ src/test/java/com/nhnacademy/marketgg/server/service/impl/DefaultCategorizationServiceTest.java
@@ -4,8 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.when;
 
 import com.nhnacademy.marketgg.server.dto.response.CategorizationRetrieveResponse;
-import com.nhnacademy.marketgg.server.repository.CategorizationRepository;
-import com.nhnacademy.marketgg.server.service.CategorizationService;
+import com.nhnacademy.marketgg.server.repository.categorization.CategorizationRepository;
+
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -13,9 +13,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.transaction.annotation.Transactional;
 
 @ExtendWith(MockitoExtension.class)
