+++ src/test/java/com/nhnacademy/marketgg/server/service/impl/DefaultCategoryServiceTest.java
@@ -17,7 +17,7 @@ import com.nhnacademy.marketgg.server.entity.Categorization;
 import com.nhnacademy.marketgg.server.entity.Category;
 import com.nhnacademy.marketgg.server.exception.categorization.CategorizationNotFoundException;
 import com.nhnacademy.marketgg.server.exception.category.CategoryNotFoundException;
-import com.nhnacademy.marketgg.server.repository.CategorizationRepository;
+import com.nhnacademy.marketgg.server.repository.categorization.CategorizationRepository;
 import com.nhnacademy.marketgg.server.repository.CategoryRepository;
 import java.util.List;
 import java.util.Optional;
